+++ webviz_config/webviz_plugin_subclasses/__init__.py
@@ -1,3 +1,2 @@
 from ._settings_group_abc import SettingsGroupABC
-from ._view_abc import ViewABC
-from ._view_element_abc import ViewElementABC
+from ._views import ViewABC, ViewElementABC, ViewLayoutElement
